get all recently viewed by user id and order by recent id desc

+++ src/main/java/com/siemensgamesa/deviation/Model/RecentlyViewedModel.java
@@ -21,7 +21,7 @@ public class RecentlyViewedModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "recentid")
-    private int recentId;
+    private long recentId;
 
     @Column(name = "userid")
     private long userId;
